Documents/OneDrive/2025/PROJEKTS/Intelekta/Tatjana-aroma/Для бота/локаольный бот/requirements.txt
aiogram==3.4.0
aiohttp==3.9.1
python-dotenv==1.0.0
gspread==6.0.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
